Improve deletion animation and prevent swiping right

+++ SwiftUIExploration/Examples/SwipeToDelete/SwipeToDeleteViewModifier.swift
@@ -26,7 +26,8 @@ struct SwipeToDeleteViewModifier: ViewModifier {
     @State private var contentSize: CGSize = .zero
     @State private var isDeletionViewShown = false
     @State private var lastOffset: CGFloat = 0 // Used when starting to drag from a visible state
-        
+    @State private var isDeleting: Bool = false
+    
     private var onDelete: () -> Void
     
     // Used to ensure that only one view is shown at any time
@@ -63,26 +64,31 @@ struct SwipeToDeleteViewModifier: ViewModifier {
             .gesture(
                 DragGesture()
                     .onChanged { value in
+                        if isDeleting {
+                            return
+                        }
                         print("View: '\(viewId)' -> lastOffset: \(lastOffset)")
                         swipeToDeleteViewId = viewId
-                        contentOffset = calculateContentOffset(translationWidth: value.translation.width)
+                        contentOffset = calculateContentOffset(translationWidth: value.translation.width, dragStartingOffset: lastOffset)
                         isDeletionViewShown = contentOffset < -minimumSwipeThreshold
-                    }
-                    .onEnded { value in
                         if contentOffset < -swipeToDeleteThreshold() {
                             deleteItem()
-                        } else {
-                            if !isDeletionViewShown {
-                                withAnimation {
-                                    contentOffset = 0
-                                }
-                            } else {
-                                withAnimation {
-                                    contentOffset = -deletionViewWidth
-                                }
-                            }
-                            lastOffset = contentOffset
                         }
+                    }
+                    .onEnded { value in
+                        if isDeleting {
+                            return
+                        }
+                        if isDeletionViewShown {
+                            withAnimation {
+                                contentOffset = -deletionViewWidth
+                            }
+                        } else {
+                            withAnimation {
+                                contentOffset = 0
+                            }
+                        }
+                        lastOffset = contentOffset
                     }
             )
     }
@@ -110,10 +116,13 @@ struct SwipeToDeleteViewModifier: ViewModifier {
     }
 
     // MARK: Private methods
-    
-    private func calculateContentOffset(translationWidth: CGFloat) -> CGFloat {
-        // TODO:  => Make a calculate offset method (that is testable)
-        return translationWidth + lastOffset
+    private func calculateContentOffset(translationWidth: CGFloat, dragStartingOffset: CGFloat) -> CGFloat {
+        // If we swipe right, stop at the edge of the view
+        if translationWidth + dragStartingOffset > 0 {
+            return 0
+        }
+        // No limit to swiping left
+        return translationWidth + dragStartingOffset
     }
     
     private func swipeToDeleteThreshold() -> CGFloat {
@@ -123,6 +132,7 @@ struct SwipeToDeleteViewModifier: ViewModifier {
     }
     
     private func deleteItem() {
+        isDeleting = true
         withAnimation {
             contentOffset = -UIScreen.main.bounds.width
             onDelete()
